feat : BaseProfileViewProtocol 생성
