+++ Terminal-iOS/Set/MyProfile/ProfileDetailProtocols.swift
@@ -8,12 +8,12 @@
 
 import UIKit
 
-protocol ProfileDetailViewProtocol: class {
+protocol ProfileDetailViewProtocol: BaseProfileViewProtocol {
     var presenter: ProfileDetailPresenterProtocol? { get set }
     
     // PRESENTER -> VIEW
-    func showUserInfo(with userInfo: UserInfo)
-    func addProjectToStackView(with project: [Project])
+//    func showUserInfo(with userInfo: UserInfo)
+//    func addProjectToStackView(with project: [Project])
 }
 
 protocol ProfileDetailWireFrameProtocol: class {
